+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [1.0.26]
 ### Fixed
 - Chickens bounding box being the wrong height
+- Fluid eggs destroying blocks
 
 ## [1.0.25] - 2022.11.04
 ### Fixed